pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hvac-climate-data-pipeline"
version = "0.1.0"
description = "A comprehensive data pipeline for processing HVAC sensor data and climate information"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Data Engineering Team", email = "team@example.com"}
]
keywords = ["hvac", "climate", "data-pipeline", "etl", "medallion-architecture"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyarrow>=12.0.0",
    "apache-airflow>=2.7.0",
    "prefect>=2.13.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "great-expectations>=0.18.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.urls]
Homepage = "https://github.com/shahabsalehi/HVAC-Climate-Data-Pipeline"
Repository = "https://github.com/shahabsalehi/HVAC-Climate-Data-Pipeline"
Documentation = "https://github.com/shahabsalehi/HVAC-Climate-Data-Pipeline/blob/main/README.md"
"Bug Tracker" = "https://github.com/shahabsalehi/HVAC-Climate-Data-Pipeline/issues"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
